Move github function from App.js
